1. Some minor fixes to the line double click mechanism 2. Change "process terminated" to "process complete" 3. Remove GTK.STOCK_NO icon on unmerged packages.
Double-clicking on the lines in the info, cautions & warnings tabs will bring the process window into focus and position the desired line in the center of the window.

Brian's sticking scroll function interferes with the correction operation of this feature.  If you move the process window to the center, this new feature works unless the line number you select is within the "sticky zone".  Brian: can you see what can be done to temporarily suspend your sticky scroll feature?