Committing the terminal process display changes I made before my computer crashed.  They have been testing as working good now, no segfaults.  The terminal now handles CR's without LF's by switching to overwrite mode. The file download progress bars are not quite right yet, I need to look for and handle the escape sequence to put the right side portion at the right side of the display screen.
A minor comment fix in version_sort.

Changed the load_web_page() to open the browser in a thread so the other windows are not frozen untill the browser is
closed.  Now the help files can remain open while still being able to work with porthole.  It also now allows for multiple
pages to be opened by porthole.